Fix Windows support (#71)
* Fix partially missing binary dependencies in setup.py

* Create and distribute Python wheel on Windows..

* Add missing Pillow Python dependency. Debug Boost search path.

* Fix ambiguous Python executable path detection for dependencies on Windows.

 * Improve cmake Python executable detection.

* Fix boost includes jiminy dependency not propagated to targets relying on it.

* Use meshcat if gepetto not available. Fix deprecated '.A' missing out.

* Fix Python binary log parser

+++ build_tools/cmake/buildPythonWheel.cmake
@@ -27,16 +27,18 @@ function(buildPythonWheel TARGET_PATH)
             COMPONENT pypi
             EXCLUDE_FROM_ALL
     )
-    install(CODE "file(REMOVE_RECURSE \"${CMAKE_BINARY_DIR}/pypi/${TARGET_NAME}/dist\")
-                  execute_process(COMMAND python setup.py sdist bdist_wheel
-                                  WORKING_DIRECTORY ${CMAKE_BINARY_DIR}/pypi/${TARGET_NAME}
-                  )
-                  file(GLOB PYTHON_WHEEL_FILENAME \"${CMAKE_BINARY_DIR}/pypi/${TARGET_NAME}/dist/*.whl\")
-                  string(REPLACE \"-linux_\" \"-manylinux1_\" PYTHON_WHEEL_FILENAME_FIXED \"\${PYTHON_WHEEL_FILENAME}\")
-                  file(RENAME \"\${PYTHON_WHEEL_FILENAME}\" \"\${PYTHON_WHEEL_FILENAME_FIXED}\")
-                  "
-            COMPONENT pypi
-            EXCLUDE_FROM_ALL
-    )
+    if(NOT WIN32)
+        install(CODE "file(REMOVE_RECURSE \"${CMAKE_BINARY_DIR}/pypi/${TARGET_NAME}/dist\")
+                      execute_process(COMMAND python setup.py sdist bdist_wheel
+                                      WORKING_DIRECTORY ${CMAKE_BINARY_DIR}/pypi/${TARGET_NAME}
+                      )
+                      file(GLOB PYTHON_WHEEL_FILENAME \"${CMAKE_BINARY_DIR}/pypi/${TARGET_NAME}/dist/*.whl\")
+                      string(REPLACE \"-linux_\" \"-manylinux1_\" PYTHON_WHEEL_FILENAME_FIXED \"\${PYTHON_WHEEL_FILENAME}\")
+                      file(RENAME \"\${PYTHON_WHEEL_FILENAME}\" \"\${PYTHON_WHEEL_FILENAME_FIXED}\")
+                      "
+                COMPONENT pypi
+                EXCLUDE_FROM_ALL
+        )
+    endif()
     set_directory_properties(PROPERTIES "${CMAKE_BINARY_DIR}/pypi/${TARGET_NAME}" ADDITIONAL_MAKE_CLEAN_FILES)
 endfunction()
